refactor: extract parseTraceFile into shared test utility, fix error path cleanup
- Move duplicated parseTraceFile/parseTraceEvents into test/lib/parse-trace-file.ts
  and update all 5 test files to use the shared utility.
- Fix error path in turbopackBuild() to abort the compilation events
  iterator and drain it before shutting down the project, preventing a
  leaked subscription on build failure.
