+++ test/e2e/app-dir/trace-build-file/trace-build-file.test.ts
@@ -1,61 +1,7 @@
 import { nextTestSetup, isNextDev, isNextStart } from 'e2e-utils'
 import { join } from 'path'
-import { existsSync, readFileSync } from 'fs'
-import type { TraceEvent } from 'next/dist/trace'
-
-interface TraceStructure {
-  events: TraceEvent[]
-  eventsByName: Map<string, TraceEvent[]>
-  eventsById: Map<string, TraceEvent>
-  rootEvents: TraceEvent[]
-  orphanedEvents: TraceEvent[]
-}
-
-function parseTraceFile(traceBuildPath: string): TraceStructure {
-  const traceContent = readFileSync(traceBuildPath, 'utf8')
-  const traceLines = traceContent
-    .trim()
-    .split('\n')
-    .filter((line) => line.trim())
-
-  const allEvents: TraceEvent[] = []
-
-  for (const line of traceLines) {
-    const events = JSON.parse(line) as TraceEvent[]
-    allEvents.push(...events)
-  }
-
-  const eventsByName = new Map<string, TraceEvent[]>()
-  const eventsById = new Map<string, TraceEvent>()
-  const rootEvents: TraceEvent[] = []
-  const orphanedEvents: TraceEvent[] = []
-
-  // Index all events
-  for (const event of allEvents) {
-    if (!eventsByName.has(event.name)) {
-      eventsByName.set(event.name, [])
-    }
-    eventsByName.get(event.name).push(event)
-    eventsById.set(event.id.toString(), event)
-  }
-
-  // Categorize events as root or orphaned
-  for (const event of allEvents) {
-    if (!event.parentId) {
-      rootEvents.push(event)
-    } else if (!eventsById.has(event.parentId.toString())) {
-      orphanedEvents.push(event)
-    }
-  }
-
-  return {
-    events: allEvents,
-    eventsByName,
-    eventsById,
-    rootEvents,
-    orphanedEvents,
-  }
-}
+import { existsSync } from 'fs'
+import { parseTraceFile } from 'parse-trace-file'
 
 describe('trace-build-file', () => {
   const { next } = nextTestSetup({
